Support Scala 2.13.x
