+++ .travis.yml
@@ -11,7 +11,8 @@ services:
 
 scala:
   - 2.11.12
-  - 2.12.8
+  - 2.12.10
+  - 2.13.1
 
 jdk: oraclejdk8
