+++ src/main/scala-2.11/org.zalando.kanadi.api/CrossVersionImports.scala
@@ -1,0 +1,5 @@
+package org.zalando.kanadi.api
+
+private[this] object CrossVersionImports
+    extends io.circe.java8.time.JavaTimeDecoders
+    with io.circe.java8.time.JavaTimeEncoders
