+++ src/main/scala-2.13/org.zalando.kanadi.api/CrossVersionImports.scala
@@ -1,0 +1,3 @@
+package org.zalando.kanadi.api
+
+private[this] object CrossVersionImports
