+++ src/main/scala/org/zalando/kanadi/api/EventTypes.scala
@@ -15,7 +15,7 @@ import com.typesafe.scalalogging.{Logger, LoggerTakingImplicit}
 import de.heikoseeberger.akkahttpcirce.ErrorAccumulatingCirceSupport._
 import enumeratum._
 import io.circe._
-import io.circe.java8.time._
+import CrossVersionImports._
 import io.circe.syntax._
 import org.mdedetrich.webmodels.{FlowId, OAuth2TokenProvider}
 import org.mdedetrich.webmodels.RequestHeaders.`X-Flow-ID`
