+++ src/test/scala/org/zalando/kanadi/CommitCursorBadResponseSpec.scala
@@ -1,28 +1,31 @@
+package org.zalando.kanadi
+
 import java.util.UUID
 
 import akka.actor.ActorSystem
 import akka.http.scaladsl.Http
 import akka.stream.ActorMaterializer
+import akka.stream.scaladsl.Sink
 import com.typesafe.config.ConfigFactory
+import io.circe.JsonObject
+import org.mdedetrich.webmodels.FlowId
 import org.specs2.Specification
 import org.specs2.concurrent.ExecutionEnv
 import org.specs2.matcher.FutureMatchers
 import org.specs2.specification.core.SpecStructure
-import org.zalando.kanadi.Config
 import org.zalando.kanadi.api._
 import org.zalando.kanadi.models.{EventTypeName, SubscriptionId}
 
 import scala.concurrent.Promise
 import scala.concurrent.duration._
-import scala.util.Success
+import scala.util._
 
-class AuthorizationSpec(implicit ec: ExecutionEnv) extends Specification with FutureMatchers with Config {
+class CommitCursorBadResponseSpec(implicit ec: ExecutionEnv) extends Specification with FutureMatchers with Config {
   override def is: SpecStructure = sequential ^ s2"""
     Create Event Type          $createEventType
-    Get Event Type             $getEventType
-    Create subscription        $createSubscription
-    Get subscription           $getSubscription
-    """
+    Create Subscription events $createSubscription
+    Stream subscription        $streamSubscriptionId
+  """
 
   val config = ConfigFactory.load()
 
@@ -46,46 +49,28 @@ class AuthorizationSpec(implicit ec: ExecutionEnv) extends Specification with Fu
   val eventsTypesClient =
     EventTypes(nakadiUri, None)
 
-  val currentSubscriptionId: Promise[SubscriptionId] = Promise()
-
-  val authorization = EventTypeAuthorization(
-    List(AuthorizationAttribute("user", "adminClientId")),
-    List(AuthorizationAttribute("user", "adminClientId")),
-    List(AuthorizationAttribute("user", "adminClientId"))
-  )
-
-  val subscriptionAuthorization = SubscriptionAuthorization(
-    List(AuthorizationAttribute("user", "adminClientId")),
-    List(AuthorizationAttribute("user", "adminClientId"))
-  )
+  val currentSubscriptionId: Promise[SubscriptionId]                             = Promise()
+  val successfullyParsedBadCommitResponse: Promise[Option[CommitCursorResponse]] = Promise()
 
   def createEventType = (name: String) => {
-    val future = eventsTypesClient.create(
-      EventType(name = eventTypeName,
-                owningApplication = OwningApplication,
-                category = Category.Business,
-                authorization = Some(authorization)))
+    val future = eventsTypesClient.create(EventType(eventTypeName, OwningApplication, Category.Business))
 
     future must be_==(()).await(retries = 3, timeout = 10 seconds)
   }
 
-  def getEventType = (name: String) => {
-    val future = eventsTypesClient.get(eventTypeName).map(_.flatMap(_.authorization))
-    future must beSome(authorization).await(retries = 3, timeout = 10 seconds)
-  }
-
   def createSubscription = (name: String) => {
+    implicit val flowId: FlowId = Utils.randomFlowId()
+    flowId.pp(name)
     val future = subscriptionsClient.createIfDoesntExist(
       Subscription(
-        id = None,
-        owningApplication = OwningApplication,
-        eventTypes = Some(List(eventTypeName)),
-        consumerGroup = Some(consumerGroup),
-        authorization = Some(subscriptionAuthorization)
+        None,
+        OwningApplication,
+        Some(List(eventTypeName)),
+        Some(consumerGroup)
       ))
 
     future.onComplete {
-      case scala.util.Success(subscription) =>
+      case Success(subscription) =>
         subscription.id.pp
         currentSubscriptionId.complete(Success(subscription.id.get))
       case _ =>
@@ -95,13 +80,30 @@ class AuthorizationSpec(implicit ec: ExecutionEnv) extends Specification with Fu
       .await(0, timeout = 5 seconds)
   }
 
-  def getSubscription = (name: String) => {
+  def streamSubscriptionId = (name: String) => {
+    implicit val flowId: FlowId = Utils.randomFlowId()
+    flowId.pp(name)
+
     val future = for {
       subscriptionId <- currentSubscriptionId.future
-      subscription   <- subscriptionsClient.get(subscriptionId)
-    } yield subscription.flatMap(_.authorization)
+      _ = subscriptionsClient.eventsStreamedSource[JsonObject](subscriptionId).map { nakadiSource =>
+        nakadiSource.source
+          .map { subscriptionEvent =>
+            subscriptionsClient
+              .commitCursors(subscriptionId, SubscriptionCursor(List(subscriptionEvent.cursor)), nakadiSource.streamId)
+              .onComplete {
+                case Success(response) =>
+                  successfullyParsedBadCommitResponse.complete(Success(response))
+                case Failure(e) =>
+                  successfullyParsedBadCommitResponse.complete(Failure(e))
+              }
 
-    future must beSome(subscriptionAuthorization).await(retries = 3, timeout = 10 seconds)
+          }
+          .runWith(Sink.foreach(_ => ()))
+      }
+      result <- successfullyParsedBadCommitResponse.future
+    } yield result
+
+    future must beSome.await(0, timeout = 2 minutes)
   }
-
 }
