+++ src/test/scala/org/zalando/kanadi/CommitCursorOmittedSpec.scala
@@ -1,3 +1,5 @@
+package org.zalando.kanadi
+
 import java.util.UUID
 
 import akka.actor.ActorSystem
@@ -11,7 +13,6 @@ import org.specs2.Specification
 import org.specs2.concurrent.ExecutionEnv
 import org.specs2.matcher.FutureMatchers
 import org.specs2.specification.core.SpecStructure
-import org.zalando.kanadi.Config
 import org.zalando.kanadi.api._
 import org.zalando.kanadi.models.{EventTypeName, SubscriptionId}
