+++ src/test/scala/org/zalando/kanadi/SubscriptionsSpec.scala
@@ -1,3 +1,5 @@
+package org.zalando.kanadi
+
 import java.util.UUID
 
 import akka.actor.ActorSystem
@@ -9,7 +11,6 @@ import org.specs2.Specification
 import org.specs2.concurrent.ExecutionEnv
 import org.specs2.specification.core.SpecStructure
 import org.specs2.specification.{AfterAll, BeforeAll}
-import org.zalando.kanadi.Config
 import org.zalando.kanadi.api.{Category, EventType, EventTypes, Events, Subscription, Subscriptions}
 import org.zalando.kanadi.models.{EventTypeName, SubscriptionId}
