+++ src/test/scala/org/zalando/kanadi/Utils.scala
@@ -1,3 +1,5 @@
+package org.zalando.kanadi
+
 import org.mdedetrich.webmodels.FlowId
 
 object Utils {
