+++ version.sbt
@@ -1,1 +1,1 @@
-version := "0.4.3-SNAPSHOT"
+version := "0.5.0"
